+++ bin/document_store.ml
@@ -48,7 +48,11 @@ let min_binding (t : t) =
       Some (path, (doc, search_results))
     )
 
-let update_content_reqs (content_reqs : Content_req_exp.t) (t : t) : t =
+let update_content_reqs
+    ~(stop_signal : Stop_signal.t)
+    (content_reqs : Content_req_exp.t)
+    (t : t)
+  : t =
   if Content_req_exp.equal content_reqs t.content_reqs then (
     t
   ) else (
@@ -56,63 +60,105 @@ let update_content_reqs (content_reqs : Content_req_exp.t) (t : t) : t =
       if Content_req_exp.is_empty content_reqs then (
         t.all_documents
       ) else (
-        String_option_map.filter_map (fun _path doc ->
-            if Index.fulfills_content_reqs content_reqs doc.Document.index then
-              Some doc
-            else
-              None
+        Eio.Fiber.first
+          (fun () ->
+             Stop_signal.await stop_signal;
+             String_option_map.empty
           )
-          t.all_documents
+          (fun () ->
+             String_option_map.filter_map (fun _path doc ->
+                 if Index.fulfills_content_reqs content_reqs doc.Document.index then
+                   Some doc
+                 else
+                   None
+               )
+               t.all_documents
+          )
       )
+    in
+    let search_results =
+      Eio.Fiber.first
+        (fun () ->
+           Stop_signal.await stop_signal;
+           String_option_map.empty
+        )
+        (fun () ->
+           String_option_map.mapi (fun _path doc ->
+               Index.search t.search_phrase doc.Document.index
+             )
+             filtered_documents)
     in
     { t with
       content_reqs;
       filtered_documents;
-      search_results =
-        String_option_map.mapi (fun _path doc ->
-            Index.search t.search_phrase doc.Document.index
-          )
-          filtered_documents;
+      search_results;
     }
   )
 
-let update_search_phrase search_phrase (t : t) : t =
+let update_search_phrase ~(stop_signal : Stop_signal.t) search_phrase (t : t) : t =
   if Search_phrase.equal search_phrase t.search_phrase then (
     t
   ) else (
+    let search_results =
+      Eio.Fiber.first
+        (fun () ->
+           Stop_signal.await stop_signal;
+           String_option_map.empty
+        )
+        (fun () ->
+           String_option_map.mapi (fun _path doc ->
+               Index.search search_phrase doc.Document.index
+             )
+             t.filtered_documents
+        )
+    in
     { t with
       search_phrase;
-      search_results =
-        String_option_map.mapi (fun _path doc ->
-            Index.search search_phrase doc.Document.index
-          )
-          t.filtered_documents;
+      search_results;
     }
   )
 
-let add_document (doc : Document.t) (t : t) : t =
+let add_document ~(stop_signal : Stop_signal.t) (doc : Document.t) (t : t) : t =
+  let filtered_documents =
+    Eio.Fiber.first
+      (fun () ->
+         Stop_signal.await stop_signal;
+         String_option_map.empty
+      )
+      (fun () ->
+         if Index.fulfills_content_reqs t.content_reqs doc.index then
+           String_option_map.add doc.path doc t.filtered_documents
+         else
+           t.filtered_documents
+      )
+  in
+  let search_results =
+    Eio.Fiber.first
+      (fun () ->
+         Stop_signal.await stop_signal;
+         String_option_map.empty
+      )
+      (fun () ->
+         String_option_map.add
+           doc.path
+           (Index.search t.search_phrase doc.index)
+           t.search_results
+      )
+  in
   { t with
     all_documents =
       String_option_map.add
         doc.path
         doc
         t.all_documents;
-    filtered_documents =
-      (if Index.fulfills_content_reqs t.content_reqs doc.index then
-         String_option_map.add doc.path doc t.filtered_documents
-       else
-         t.filtered_documents
-      );
-    search_results =
-      String_option_map.add
-        doc.path
-        (Index.search t.search_phrase doc.index)
-        t.search_results;
+    filtered_documents;
+    search_results;
   }
 
 let of_seq (s : Document.t Seq.t) =
+  let dummy_stop_signal = Stop_signal.make () in
   Seq.fold_left (fun t doc ->
-      add_document doc t
+      add_document ~stop_signal:dummy_stop_signal doc t
     )
     empty
     s