+++ bin/stop_signal.mli
@@ -1,0 +1,7 @@
+type t
+
+val make : unit -> t
+
+val signal : t -> unit
+
+val await : t -> unit